pass: add --generate-autofill option
This generates the mapping for the autofill.js plugin
automatically from your data. It's a little weird to keep
the URL field inside the pass data, since it isn't a secret
(after all, we copy it out to pass.js here). But it's
convenient to keep it contained with the password data it
matches.