autofill: make end-of-pattern optional
When given a URL like "https://example.com", we look for
end-of-string, "/", or "?" at the end to avoid mismatching
"https://example.com.other.domain". However, if the URL we
get already has a delimiting character like "/", we can just
leave it be.

This would be much simpler if just matched the domain part
by itself, but I actually have some cases that differentiate
based on the path. That may not be worth supporting, though
(existing password managers usually don't).