Revert "autofill: recognize "login" as a login box"
This reverts commit 9636e3cab3921f9ab243a8aa404da04a4097b73d.

That commit improved matching on some sites, but makes other
sites much worse (e.g., ones that have "Login$Username",
"Login$Submit", and so on). So clearly we need smarter
heuristics. It may be enough to just set the first such
element we see.

Unfortunately, I don't recall the exact site which prompted
9636e3cab in the first place, so I don't have data with
which to make a heuristic. Let's back it out for now, and we
can replace it with a better rule when the problem
reappears.