autofill dynamically from perl script
The original idea behind --generate-autofill is that we
would generate host-matches via regex, so that we could know
which candidates we have for auto-filling even before
opening the password file. The advantage being that we would
not ask to unlock the file only to discover there was
nothing interesting.

However, since the pentadactyl side of the autofill does not
run automatically, but rather only when the user requests
it, this isn't much of an advantage. The user doesn't bother
triggering it unless they expect there to be a fillable
item.

Therefore we can push the filling logic down into the perl
script without any significant loss, and several advantages:

  1. You no longer have to pre-generate the autofill
     javascript when the set of autofills changes.

  2. The quoted regexes were ugly and hard to debug.

  3. Better privacy; the set of regexes is not in cleartext
     on the disk any more.

There is one case it does not handle as well: we used to
accept a regex as a "fillurl", which meant that a given
entry could match multiple sites. It would be possible to
support this by allowing multiple "url" fields for each
element, but this patch does not do that yet.